atualizacao do plano

+++ app/Http/Controllers/Admin/PlanController.php
@@ -46,10 +46,19 @@ class PlanController extends Controller
     }
 
     public function destroy($url) {
-        $plan = $this->plan->where('url', $url)->first();
+        $plan = $this->plan
+                        ->with('details')
+                        ->where('url', $url)
+                        ->first();
 
         if (!$plan) {
             return redirect()->back();
+        }
+
+        if($plan->details->count() > 0) {
+            return redirect()
+                        ->back()
+                        ->with('error', 'Existem detalhes vinculados a este plano');
         }
 
         $plan->delete();
